requirements.txt
requests
restrictedpython
pydantic
pyyaml
nats-py
mako
python-dotenv
pytest
pytest-asyncio